refactor: put DistoSyncCommand into its own file
